Refactor date formatting in AnnouncementItem component
- Moved the date formatting logic to a separate utility function in `utils/dateUtils` for better organization and reusability.
- Removed the inline date formatting function and updated the component to import the new utility, enhancing code clarity and maintainability.
